Move request bonus logic to notifs icon
